Fix build errors

+++ src/bsp/raspi/video.rs
@@ -2,11 +2,12 @@
 // Access function for device
 // Font renderring/Rendering Characters
 
+#[rustfmt::skip]
 pub(super) mod VIDEO {
   pub const COLOR_DEPTH: u8 = 8;
 
   pub mod RESOLUTION {
-    pub const HEIGHT: u8 = 1024;
-    pub const WIDTH:  u8 = 768;
+    pub const HEIGHT: u16 = 1024;
+    pub const WIDTH:  u16 = 768;
   }
 }
